Reorganize utilities
This commit splits the utilities across several subpackages in order to
improve their organization.  This will hopefully make it easier for
utilizers to find the utility they need.
