+++ utils/locations/files_test.go
@@ -12,7 +12,7 @@
 // implied.  See the License for the specific language governing
 // permissions and limitations under the License.
 
-package utils
+package locations
 
 import (
 	"testing"
@@ -368,153 +368,5 @@ func TestFileLocationIncrFormFeedBeginningOfLine(t *testing.T) {
 		File: "file",
 		B:    FilePos{3, 1},
 		E:    FilePos{3, 2},
-	}, loc)
-}
-
-func TestArgLocationImplementsLocation(t *testing.T) {
-	assert.Implements(t, (*common.Location)(nil), &ArgLocation{})
-}
-
-func TestArgLocationString0Columns(t *testing.T) {
-	loc := ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 3},
-	}
-
-	result := loc.String()
-
-	assert.Equal(t, "1:3", result)
-}
-
-func TestArgLocationString1Columns(t *testing.T) {
-	loc := ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 4},
-	}
-
-	result := loc.String()
-
-	assert.Equal(t, "1:3", result)
-}
-
-func TestArgLocationString2Columns(t *testing.T) {
-	loc := ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 5},
-	}
-
-	result := loc.String()
-
-	assert.Equal(t, "1:3-5", result)
-}
-
-func TestArgLocationString2Arguments(t *testing.T) {
-	loc := ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{2, 1},
-	}
-
-	result := loc.String()
-
-	assert.Equal(t, "1:3-2:1", result)
-}
-
-func TestArgLocationThruBase(t *testing.T) {
-	loc1 := ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 5},
-	}
-	loc2 := ArgLocation{
-		B: ArgPos{2, 2},
-		E: ArgPos{2, 4},
-	}
-
-	result, err := loc1.Thru(loc2)
-
-	assert.NoError(t, err)
-	assert.Equal(t, ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{2, 2},
-	}, result)
-}
-
-func TestArgLocationThruSplitLocation(t *testing.T) {
-	loc1 := ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 5},
-	}
-	loc2 := &common.MockLocation{}
-
-	result, err := loc1.Thru(loc2)
-
-	assert.Same(t, common.ErrSplitLocation, err)
-	assert.Nil(t, result)
-}
-
-func TestArgLocationThruEndBase(t *testing.T) {
-	loc1 := ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 5},
-	}
-	loc2 := ArgLocation{
-		B: ArgPos{2, 2},
-		E: ArgPos{2, 4},
-	}
-
-	result, err := loc1.ThruEnd(loc2)
-
-	assert.NoError(t, err)
-	assert.Equal(t, ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{2, 4},
-	}, result)
-}
-
-func TestArgLocationThruEndSplitLocation(t *testing.T) {
-	loc1 := ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 5},
-	}
-	loc2 := &common.MockLocation{}
-
-	result, err := loc1.ThruEnd(loc2)
-
-	assert.Same(t, common.ErrSplitLocation, err)
-	assert.Nil(t, result)
-}
-
-func TestArgLocationIncrBase(t *testing.T) {
-	loc := ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 5},
-	}
-
-	result := loc.Incr('t', 8)
-
-	assert.Equal(t, ArgLocation{
-		B: ArgPos{1, 5},
-		E: ArgPos{1, 6},
-	}, result)
-	assert.Equal(t, ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 5},
-	}, loc)
-}
-
-func TestArgLocationIncrEOF(t *testing.T) {
-	loc := ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 5},
-	}
-
-	result := loc.Incr(common.EOF, 8)
-
-	assert.Equal(t, ArgLocation{
-		B: ArgPos{1, 5},
-		E: ArgPos{1, 5},
-	}, result)
-	assert.Equal(t, ArgLocation{
-		B: ArgPos{1, 3},
-		E: ArgPos{1, 5},
 	}, loc)
 }
