+++ utils/nodes/visualize.go
@@ -12,7 +12,7 @@
 // implied.  See the License for the specific language governing
 // permissions and limitations under the License.
 
-package utils
+package nodes
 
 import (
 	"bytes"
