+++ utils/tokenstreams/channels_test.go
@@ -12,7 +12,7 @@
 // implied.  See the License for the specific language governing
 // permissions and limitations under the License.
 
-package utils
+package tokenstreams
 
 import (
 	"testing"
@@ -78,82 +78,4 @@ func TestChanTokenStreamDone(t *testing.T) {
 
 	_, ok := <-obj.Chan
 	assert.False(t, ok)
-}
-
-func TestListTokenStreamImplementsTokenStream(t *testing.T) {
-	assert.Implements(t, (*common.TokenStream)(nil), &listTokenStream{})
-}
-
-func TestNewListTokenStream(t *testing.T) {
-	toks := []*common.Token{{}, {}, {}}
-
-	result := NewListTokenStream(toks)
-
-	assert.Equal(t, &listTokenStream{
-		toks: toks,
-	}, result)
-}
-
-func TestListTokenStreamNextUnstarted(t *testing.T) {
-	toks := []*common.Token{{}, {}, {}}
-	obj := &listTokenStream{
-		toks: toks,
-	}
-
-	result := obj.Next()
-
-	assert.Same(t, toks[0], result)
-	assert.Equal(t, &listTokenStream{
-		toks:    toks,
-		started: true,
-	}, obj)
-}
-
-func TestListTokenStreamNextStarted(t *testing.T) {
-	toks := []*common.Token{{}, {}, {}}
-	obj := &listTokenStream{
-		toks:    toks,
-		started: true,
-	}
-
-	result := obj.Next()
-
-	assert.Same(t, toks[1], result)
-	assert.Equal(t, &listTokenStream{
-		toks:    toks,
-		idx:     1,
-		started: true,
-	}, obj)
-}
-
-func TestListTokenStreamNextEnding(t *testing.T) {
-	toks := []*common.Token{{}, {}, {}}
-	obj := &listTokenStream{
-		toks:    toks,
-		idx:     2,
-		started: true,
-	}
-
-	result := obj.Next()
-
-	assert.Nil(t, result)
-	assert.Equal(t, &listTokenStream{
-		toks:    toks,
-		idx:     2,
-		started: true,
-	}, obj)
-}
-
-func TestNewAsyncTokenStream(t *testing.T) {
-	toks := []*common.Token{{}, {}, {}}
-	ts := NewListTokenStream(toks)
-
-	result := NewAsyncTokenStream(ts)
-
-	i := 0
-	for tok := result.Next(); tok != nil; tok = result.Next() {
-		assert.Same(t, toks[i], tok)
-		i++
-	}
-	assert.Equal(t, len(toks), i)
 }
